feat: rerun OCR from upload preview
